+++ src/Widget/CheckboxList/CheckboxItem.php
@@ -17,6 +17,7 @@ final class CheckboxItem
         public string $label,
         public bool $encodeLabel,
         public array $labelAttributes = [],
+        public bool $labelWrap = true,
     ) {
     }
 }